+++ Plugin~/WebRTCPluginTest/ContextTest.cpp
@@ -1,9 +1,8 @@
 #include "pch.h"
+
 #include "GraphicsDeviceTestBase.h"
 #include "GraphicsDevice/IGraphicsDevice.h"
 #include "GraphicsDevice/ITexture2D.h"
-#include "Codec/EncoderFactory.h"
-#include "Codec/IEncoder.h"
 #include "Context.h"
 
 #include "rtc_base/ref_counted_object.h"
@@ -13,25 +12,24 @@ namespace unity
 namespace webrtc
 {
 
+using namespace ::webrtc;
+
 class ContextTest : public GraphicsDeviceTestBase
 {
 protected:
     const int width = 256;
     const int height = 256;
     std::unique_ptr<Context> context;
-    std::unique_ptr<IEncoder> encoder_;
     DelegateVideoFrameResize callback_videoframeresize;
 
     void SetUp() override {
         GraphicsDeviceTestBase::SetUp();
-        EXPECT_NE(nullptr, m_device);
-
-        encoder_ = EncoderFactory::GetInstance().Init(width, height, m_device, m_encoderType, m_textureFormat);
-        EXPECT_NE(nullptr, encoder_);
+        EXPECT_NE(nullptr, device());
 
         context = std::make_unique<Context>();
         callback_videoframeresize = &OnFrameSizeChange;
     }
+
     void TearDown() override {
         GraphicsDeviceTestBase::TearDown();
     }
@@ -41,11 +39,11 @@ protected:
     }
 };
 TEST_P(ContextTest, InitializeAndFinalizeEncoder) {
-    const std::unique_ptr<ITexture2D> tex(m_device->CreateDefaultTextureV(width, height, m_textureFormat));
+    const std::unique_ptr<ITexture2D> tex(device()->CreateDefaultTextureV(width, height, m_textureFormat));
     EXPECT_NE(nullptr, tex);
     const auto source = context->CreateVideoSource();
     const auto track = context->CreateVideoTrack("video", source);
-    EXPECT_TRUE(context->InitializeEncoder(encoder_.get(), track));
+
     context->RemoveRefPtr(track);
     context->RemoveRefPtr(source);
 }
@@ -57,12 +55,12 @@ TEST_P(ContextTest, CreateAndDeleteMediaStream) {
 
 
 TEST_P(ContextTest, CreateAndDeleteVideoTrack) {
-    const std::unique_ptr<ITexture2D> tex(m_device->CreateDefaultTextureV(width, height, m_textureFormat));
+    const std::unique_ptr<ITexture2D> tex(device()->CreateDefaultTextureV(width, height, m_textureFormat));
     EXPECT_NE(nullptr, tex.get());
     const auto source = context->CreateVideoSource();
     const auto track = context->CreateVideoTrack("video", source);
     EXPECT_NE(nullptr, track);
-    EXPECT_TRUE(context->InitializeEncoder(encoder_.get(), track));
+
     context->RemoveRefPtr(track);
     context->RemoveRefPtr(source);
 }
@@ -87,7 +85,7 @@ TEST_P(ContextTest, AddAndRemoveAudioTrackToMediaStream) {
 }
 
 TEST_P(ContextTest, AddAndRemoveVideoTrackToMediaStream) {
-    const std::unique_ptr<ITexture2D> tex(m_device->CreateDefaultTextureV(width, height, m_textureFormat));
+    const std::unique_ptr<ITexture2D> tex(device()->CreateDefaultTextureV(width, height, m_textureFormat));
     const auto stream = context->CreateMediaStream("videostream");
     const auto source = context->CreateVideoSource();
     const auto track = context->CreateVideoTrack("video", source);
@@ -131,7 +129,7 @@ TEST_P(ContextTest, EqualRendererGetById) {
 }
 
 TEST_P(ContextTest, AddAndRemoveVideoRendererToVideoTrack) {
-    const std::unique_ptr<ITexture2D> tex(m_device->CreateDefaultTextureV(width, height, m_textureFormat));
+    const std::unique_ptr<ITexture2D> tex(device()->CreateDefaultTextureV(width, height, m_textureFormat));
     const auto source = context->CreateVideoSource();
     const auto track = context->CreateVideoTrack("video", source);
     const auto renderer = context->CreateVideoRenderer(callback_videoframeresize, true);
@@ -142,7 +140,7 @@ TEST_P(ContextTest, AddAndRemoveVideoRendererToVideoTrack) {
     context->RemoveRefPtr(source);
 }
 
-INSTANTIATE_TEST_SUITE_P(GraphicsDeviceParameters, ContextTest, testing::ValuesIn(VALUES_TEST_ENV));
+INSTANTIATE_TEST_SUITE_P(GfxDeviceAndColorSpece, ContextTest, testing::ValuesIn(VALUES_TEST_ENV));
 
 } // end namespace webrtc
 } // end namespace unity
